feat: Add CreateFeed view for creating new posts
The `CreateFeed` view is added to allow users to create new posts. It includes a form for entering the post text and uploading an image. The form is submitted via a POST request and the post is associated with the currently logged-in user as its author. After successful submission, the user is redirected to the home page.

This commit addresses the need to provide a way for users to create new posts and contributes to the overall functionality of the application.